refactor: 라인 추가 요금 등록 시 요금 타입 및 메서드 네이밍 변경

+++ src/main/java/nextstep/subway/domain/Line.java
@@ -65,7 +65,7 @@ public class Line {
         sections.delete(station);
     }
 
-    public void updateExtraFare(int extraFare) {
+    public void updateExtraFare(long extraFare) {
         this.extraFare = extraFare;
     }
